+++ src/App.js
@@ -2,23 +2,12 @@ import logo from "./logo.svg";
 import "./App.css";
 import React, { useState } from "react";
 import { ReactDOM } from "react";
-
-const useInput = (initialValue) => {
-  const [value, setValue] = useState(initialValue);
-  const onChange = event => {
-    console.log(event.target);
-  }
-  return { value, onChange} ;
-};
+import HookuseInput from "./HookuseInput";
+import HookuseState from "./HookuseState";
 
 function App() {
-  const name = useInput("Mr.");
   return (
-    <div className="App">
-      <h1>hello</h1>
-      <input placeholder="Name" {...name}></input> 
-      {/* spread 연산자 활용 */}
-    </div>
+    <HookuseState></HookuseState>
   );
 }
