+++ src/HookuseState.js
@@ -1,0 +1,17 @@
+import React, { useState } from "react";
+
+function HookuseState() {
+  const [item, setItem] = useState(1);
+  const incrementItem = () => setItem(item + 1);
+  const decrementItem = () => setItem(item - 1);
+  return (
+    <div className="App">
+      <h1>hello {item}</h1>
+      <h2>Start edting to see some magic happen!</h2>
+      <button onClick={incrementItem}>increment</button>
+      <button onClick={decrementItem}>decrement</button>
+    </div>
+  );
+}
+
+export default HookuseState;
